requirements.txt
# Core dependencies
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
google-cloud-secret-manager==2.18.0
python-dotenv==1.0.0

# LLM integration
openai==1.51.0
anthropic==0.7.8
httpx==0.27.0

# Email processing
beautifulsoup4==4.12.2
lxml==4.9.3

# Web framework for Cloud Run
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Utilities
pydantic==2.5.0
pydantic-settings==2.1.0
schedule==1.2.0
loguru==0.7.2
tenacity==8.2.3

# Development
pytest==7.4.3
black==23.11.0
flake8==6.1.0
